Moved collision algorithms to separate file
