+++ src/amuse/community/huayno/src/Makefile
@@ -26,7 +26,8 @@ RM = rm
 
 OBJS = evolve.o evolve_shared.o evolve_sf.o evolve_cc.o \
   evolve_ok.o evolve_kepler.o universal_variable_kepler.o evolve_bs.o \
-  evolve_shared_collisions.o evolve_error_control.o simple_map.o simple_hash.o
+  evolve_shared_collisions.o evolve_error_control.o simple_map.o simple_hash.o \
+  collision_check.o
 
 all: libhuayno.a
