+++ src/amuse/community/huayno/src/collision_check.h
@@ -1,0 +1,6 @@
+#ifndef COLLISION_CHECK_H
+#define COLLISION_CHECK_H
+
+void detect_collisions(struct sys s);
+
+#endif // COLLISION_CHECK_H
